+++ src/main/java/com/dyz/authentication/service/impl/SysMenuServiceImpl.java
@@ -3,8 +3,8 @@ package com.dyz.authentication.service.impl;
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
 import com.dyz.authentication.entity.*;
-import com.dyz.authentication.entity.VO.SysMenuVo;
-import com.dyz.authentication.entity.VO.SysRoleAndPermissionVo;
+import com.dyz.authentication.entity.Vo.SysMenuVo;
+import com.dyz.authentication.entity.Vo.SysRoleAndPermissionVo;
 import com.dyz.authentication.mapper.SysMenuMapper;
 import com.dyz.authentication.mapper.SysRoleMenuMapper;
 import com.dyz.authentication.mapper.SysUserMapper;
@@ -14,11 +14,7 @@ import com.dyz.authentication.service.SysRoleMenuService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.ArrayList;
-import java.util.Comparator;
-import java.util.List;
-import java.util.Map;
+import java.util.*;
 import java.util.stream.Collectors;
 
 /**
@@ -51,18 +47,48 @@ public class SysMenuServiceImpl extends ServiceImpl<SysMenuMapper, SysMenu> impl
      * @return java.util.List<com.dyz.authentication.entity.SysMenu>
      **/
     @Override
-    public List<SysMenu> getMenuListByUserName(String username) {
+    public List<SysMenuVo> getMenuListByUserName(String username) {
+        long currentTimeMillis = System.currentTimeMillis();
+        //  最终结果集合
+        List<SysMenuVo> resultList = new ArrayList<>();
+        // 将menu对象转换为menuVo对象的集合
+        List<SysMenuVo> voList = new ArrayList<>();
+        // 所有菜单集合
         List<SysMenu> menuList = new ArrayList<>();
+        // 去重后的集合
+        List<SysMenu> menus = new ArrayList<>();
         SysUser user = userMapper.selectOne(new QueryWrapper<SysUser>().eq("user_name", username));
         List<SysUserRole> userRoleList = userRoleMapper.selectList(new QueryWrapper<SysUserRole>().eq("user_id", user.getId()));
-        for (SysUserRole userRole : userRoleList) {
-            List<SysRoleMenu> roleMenuList = roleMenuMapper.selectList(new QueryWrapper<SysRoleMenu>().eq("role_id", userRole.getRoleId()));
-            for (SysRoleMenu roleMenu : roleMenuList) {
-                SysMenu menu = menuMapper.selectOne(new QueryWrapper<SysMenu>().eq("id", roleMenu.getMenuId()));
-                menuList.add(menu);
+        userRoleList.stream().parallel().map(
+                userRole -> roleMenuMapper.selectList(new QueryWrapper<SysRoleMenu>().eq("role_id", userRole.getRoleId())))
+                .forEach(roleMenuList -> roleMenuList.stream().parallel().forEach(roleMenu -> {
+            SysMenu menu = menuMapper.selectOne(new QueryWrapper<SysMenu>().eq("id", roleMenu.getMenuId()));
+            menuList.add(menu);
+        }));
+//        menuList = menuMapper.getMenuListByUserName(username);
+        menus = menuList.stream().parallel().collect(Collectors.collectingAndThen(
+                Collectors.toCollection(() -> new TreeSet<>(Comparator.comparing(SysMenu::getMenuId))),
+                ArrayList::new));
+        for (SysMenu menu : menus) {
+            SysMenuVo sysMenuVo = new SysMenuVo();
+            voList.add(menuToMenuVo(sysMenuVo, menu));
+        }
+        // 根据父级Id进行分组
+        Map<String, List<SysMenuVo>> map = voList.stream().parallel().filter(item -> item.getParentId() != null).collect(Collectors.groupingBy(SysMenu::getParentId));
+        // 父级菜单集合
+        List<SysMenuVo> collect = voList.stream().parallel().filter(item -> item.getParentId() == null).collect(Collectors.toList());
+        if (map.size() <= collect.size()) {
+            for (SysMenuVo vo : collect) {
+                for (String s : map.keySet()) {
+                    if (vo.getMenuId().equals(s)) {
+                        List<SysMenuVo> list = map.get(s).stream().parallel().sorted(Comparator.comparing(SysMenu::getMenuSort)).collect(Collectors.toList());
+                        vo.setMenuVoList(list);
+                    }
+                }
+                resultList.add(vo);
             }
         }
-        return menuList;
+        return resultList.stream().sorted(Comparator.comparing(SysMenu::getMenuSort)).collect(Collectors.toList());
     }
 
     /*
@@ -76,16 +102,7 @@ public class SysMenuServiceImpl extends ServiceImpl<SysMenuMapper, SysMenu> impl
         List<SysMenu> list = this.list();
         for (SysMenu sysMenu : list) {
             SysMenuVo sysMenuVo = new SysMenuVo();
-            sysMenuVo.setMenuId(sysMenu.getMenuId());
-            sysMenuVo.setParentName(null);
-            sysMenuVo.setCreateTime(sysMenu.getCreateTime());
-            sysMenuVo.setUpdateTime(sysMenu.getUpdateTime());
-            sysMenuVo.setMenuUrl(sysMenu.getMenuUrl());
-            sysMenuVo.setMenuSort(sysMenu.getMenuSort());
-            sysMenuVo.setDescription(sysMenu.getDescription());
-            sysMenuVo.setParentId(sysMenu.getParentId());
-            sysMenuVo.setMenuName(sysMenu.getMenuName());
-            resultList.add(sysMenuVo);
+            resultList.add(menuToMenuVo(sysMenuVo, sysMenu));
         }
         for (SysMenuVo sysMenuVo : resultList) {
             for (SysMenu menu : list) {
@@ -107,60 +124,53 @@ public class SysMenuServiceImpl extends ServiceImpl<SysMenuMapper, SysMenu> impl
      **/
     @Override
     public List<SysRoleAndPermissionVo> getMenuListByRoleId(String roleId) {
-        // 存储结果的集合
-        List<SysRoleAndPermissionVo> resultList = new ArrayList<>();
-        // 存储menu转换为SysRoleAndPermissionVo集合
-        List<SysRoleAndPermissionVo> permissionVoList = new ArrayList<>();
         // 菜单集合
         List<SysMenu> menuList = this.list();
         // 角色菜单集合
         List<SysRoleMenu> roleMenus = roleMenuMapper.selectList(new QueryWrapper<SysRoleMenu>().eq("role_id", roleId));
-        // 将Menu对象集合转换为SysRoleAndPermissionVo对象集合
-        for (SysMenu menu : menuList) {
-            SysRoleAndPermissionVo permissionVo = new SysRoleAndPermissionVo(menu.getMenuId(), menu.getMenuName(), menu.getParentId());
-            permissionVoList.add(permissionVo);
-        }
-        // 给当前角色权限范围内的菜单对象赋值roleId
-        for (SysRoleAndPermissionVo permissionVo : permissionVoList) {
-            for (SysRoleMenu roleMenu : roleMenus) {
-                if (permissionVo.getId().equals(roleMenu.getMenuId())){
-                    permissionVo.setRoleId(roleMenu.getRoleId());
-                }
-            }
-        }
-        // 父级菜单集合
-        List<SysRoleAndPermissionVo> collect = permissionVoList.stream().filter(item -> item.getPid() == null).collect(Collectors.toList());
-        // 根据parentId分类
-        Map<String, List<SysRoleAndPermissionVo>> map = permissionVoList.stream().filter(item -> item.getPid() != null).collect(Collectors.groupingBy(SysRoleAndPermissionVo::getPid));
-        if (collect.size() == map.size()){
-            for (SysRoleAndPermissionVo vo : collect) {
-                for (String s : map.keySet()) {
-                    if (vo.getId().equals(s)){
-                        vo.setChildrenList(map.get(s));
-                    }
-                }
-                resultList.add(vo);
-            }
-        }
-        return resultList;
+        return ListToHierarchyListService.listToHierarchyList(menuList,roleMenus);
     }
-    
+
     /*
      * @Description 删除菜单
-     * @param ids 
+     * @param ids
      * @return boolean
      **/
     @Override
     @Transactional
     public boolean delete(List<String> ids) {
-        if (ids != null){
+        if (ids != null) {
             for (String id : ids) {
-                boolean remove = roleMenuService.remove(new QueryWrapper<SysRoleMenu>().eq("menu_id", id));
-                if (remove != true); return false;
+                List<SysRoleMenu> list = roleMenuService.list(new QueryWrapper<SysRoleMenu>().eq("menu_id", id));
+                if (list.size() == 0) {
+                    continue;
+                } else {
+                    roleMenuService.remove(new QueryWrapper<SysRoleMenu>().eq("menu_id", id));
+                }
             }
             boolean b = this.removeByIds(ids);
-            if (b); return true;
+            if (b) ;
+            return true;
         }
         return false;
+    }
+
+    /*
+     * @Description menu对象转为menuvo对象
+     * @param menuVo
+     * @param menu
+     **/
+    public SysMenuVo menuToMenuVo(SysMenuVo menuVo, SysMenu menu) {
+        menuVo.setMenuId(menu.getMenuId());
+        menuVo.setParentName(null);
+        menuVo.setMenuVoList(null);
+        menuVo.setCreateTime(menu.getCreateTime());
+        menuVo.setUpdateTime(menu.getUpdateTime());
+        menuVo.setMenuUrl(menu.getMenuUrl());
+        menuVo.setMenuSort(menu.getMenuSort());
+        menuVo.setDescription(menu.getDescription());
+        menuVo.setParentId(menu.getParentId());
+        menuVo.setMenuName(menu.getMenuName());
+        return menuVo;
     }
 }
